features/favourites_list/src/main/java/io/audioshinigami/favourites_list/adaptor/CharacterFavouriteTouchHelper.kt
/*
 * MIT License
 *
 * Copyright (c) 8/11/2020 13:1   David Osemwota.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NON INFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package io.audioshinigami.favourites_list.adaptor

import androidx.recyclerview.widget.ItemTouchHelper
import androidx.recyclerview.widget.ItemTouchHelper.ACTION_STATE_IDLE
import androidx.recyclerview.widget.ItemTouchHelper.LEFT
import androidx.recyclerview.widget.ItemTouchHelper.RIGHT
import androidx.recyclerview.widget.RecyclerView

/**
 * Helper class for [RecyclerView] viewHolder which allows to swipe.
 *
 * Implements @see [ItemTouchHelper.SimpleCallback]
 */
class CharacterFavouriteTouchHelper(
        private val onSwiped: (Int) -> Unit
): ItemTouchHelper.SimpleCallback(
        ACTION_STATE_IDLE,
        LEFT or RIGHT
) {

    /**
     * Called when ItemTouchHelper wants to move from old position to new position.
     */
    override fun onMove(
            recyclerView: RecyclerView,
            viewHolder: RecyclerView.ViewHolder,
            target: RecyclerView.ViewHolder): Boolean = true

    /**
     * Called when viewHolder item is swiped.
     *
     * @param viewHolder The viewHolder which is swiped.
     * @param direction The direction swiped . eg LEFT, RIGHT , UP or DOWN.
     */
    override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
        onSwiped(viewHolder.adapterPosition)
    }
}